+++ examples/cat.rs
@@ -1,6 +1,3 @@
-extern crate fatfs;
-extern crate fscommon;
-
 use std::env;
 use std::fs::File;
 use std::io::{self, prelude::*};